requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pypdf
sentence-transformers
faiss-cpu
numpy
requests
openai
torch
pinecone

